requirements.txt
aiosqlite
discord.py
git+https://github.com/Rapptz/discord-ext-menus
python-dotenv